Transition to using uart class as uart driver
- Deleted uart0.cpp/hpp and uart2.cpp/hpp in place of uart.hpp
- Added a single global uart0 object for use anywhere in the code.
- newlib and startup code now uses uart0 object.
- Created a global.cpp/hpp file to hold the uart0 global object as well
  as the other global objects held in startup.hpp.
- startup.hpp was removed in place of globals.cpp
- Without startup.hpp, interrupts were moved to the interrupts.cpp/hpp
  which involved changing each include of startup.hpp to interrupts.hpp
