+++ firmware/examples/I2cDevice/source/main.cpp
@@ -1,6 +1,6 @@
 #include <inttypes.h>
 #include <cstdint>
-#include "L0_LowLevel/startup.hpp"
+#include "L0_LowLevel/interrupt.hpp"
 #include "L1_Drivers/i2c.hpp"
 #include "L2_Utilities/log.hpp"
 #include "L3_HAL/device_memory_map.hpp"
