+++ firmware/examples/SystemTimer/source/low_level_init.cpp
@@ -1,9 +1,10 @@
 #include <cstdint>
 
+#include "L0_LowLevel/interrupt.hpp"
 #include "L0_LowLevel/LPC40xx.h"
-#include "L0_LowLevel/startup.hpp"
 #include "L1_Drivers/pin.hpp"
 #include "L2_Utilities/log.hpp"
+#include "L4_Application/globals.hpp"
 
 constexpr uint8_t kGpioFunction = 0b000;
 Pin led3(Pin::CreatePin<1, 15>());
