+++ firmware/examples/TaskScheduler/source/main.cpp
@@ -1,6 +1,5 @@
 #include <cstdarg>
 
-#include "L0_LowLevel/uart0.hpp"
 #include "L2_Utilities/log.hpp"
 #include "L4_Application/task.hpp"
