+++ firmware/examples/Uart/source/main.cpp
@@ -1,43 +1,36 @@
 #include <project_config.hpp>
 
-#include <cstdint>
 #include <cstdio>
 
 #include "L0_LowLevel/delay.hpp"
-#include "L0_LowLevel/LPC40xx.h"
 #include "L1_Drivers/uart.hpp"
-#include "L2_Utilities/debug.hpp"
+#include "L2_Utilities/log.hpp"
 
 int main(void)
 {
-    // This example code uses a loopback test. Please connect the
-    // relevant TxD and RxD pins together. For this example please
-    // connect the TxD2 and RxD2 pins together
-
-    Uart test(Uart::Channels::kUart2);
-    char receive;
-    char input2 [] = "Hello World";
-
+    // This example code uses a loopback test Please connect the relevant TxD
+    // and RxD pins together. For this example please connect the TxD2 and RxD2
+    // pins together.
+    Uart uart2(Uart::Channels::kUart2);
     // Calls the initialize function and checks if the
     // function fails due to wrong input.
-    if (test.Initialize(38400) == 0)
-    {
-        printf("Fail!");
-    }
-
+    uart2.Initialize(38400);
     // While loop to constantly send the array of data
-    while (1)
+    while (true)
     {
+        char message[] = "Hello World!";
         // This first loop sends all the data within the array
-        for (int i=0; input2[i] != '\0'; i++) test.Send(input2[i]);
-
-        // This second loop receives all the sent data and prints it out
-        for (int j=0; input2[j] != '\0'; j++)  // 12
+        for (size_t i = 0; message[i] != '\0'; i++)
         {
-            receive = test.Receive();
-            printf("%c", receive);
+          uart2.Send(message[i]);
         }
-            printf("\n");
+        // This second loop receives all the sent data and prints it out
+        for (size_t i = 0; i < SJ2_ARRAY_LENGTH(message)-1; i++)
+        {
+            char receive = uart2.Receive();
+            putchar(receive);
+        }
+        putchar('\n');
         Delay(1000);
     }
     return 0;
